refactor: update stripe payment flow
