+++ App/utils/index.ts
@@ -1,5 +1,5 @@
 // Utilities entrypoint intentionally left minimal.
 export * from './userProfile';
-export * from './startPaymentFlow';
+export * from '../hooks/usePaymentFlow';
 export * from './transactionLogger';
